[tracer] Refactor code into its own module.
