+++ lib/apis/gravity.js
@@ -2,6 +2,7 @@
 import { assign, omit } from "lodash"
 import fetch from "./fetch"
 import config from "config"
+import { headers as requestIDHeaders } from "../requestIDs"
 
 const { GRAVITY_API_BASE } = process.env
 
@@ -12,11 +13,7 @@ export default (path, accessToken, fetchOptions = {}) => {
   const requestIDs = fetchOptions.requestIDs
   if (requestIDs) {
     fetchParams = omit(fetchOptions, "requestIDs")
-    assign(headers, {
-      "X-Request-Id": requestIDs.requestID,
-      "x-datadog-trace-id": requestIDs.traceId,
-      "x-datadog-parent-id": requestIDs.parentSpanId,
-    })
+    assign(headers, requestIDHeaders(requestIDs))
   }
 
   if (accessToken) assign(headers, { "X-ACCESS-TOKEN": accessToken })
