+++ lib/mergeSchemas.js
@@ -42,7 +42,7 @@ function createConvectionLink() {
   return middlewareLink.concat(authMiddleware).concat(httpLink)
 }
 
-export default async function mergeSchemas() {
+export async function mergeSchemas() {
   const convectionLink = createConvectionLink()
 
   const convectionSchema = await makeRemoteExecutableSchema({
